Finalized code for Flask-JWT-Extended.

+++ sections_flask-jwt-extended/resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt_claims, fresh_jwt_required
+from flask_jwt_extended import get_jwt_identity, jwt_required, get_jwt_claims, fresh_jwt_required, jwt_optional
 from models.item import ItemModel
 
 
@@ -35,33 +35,38 @@ class Item(Resource):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
 
-        data = Item.parser.parse_args()
+        data = self.parser.parse_args()
 
-        item = ItemModel(name, data['price'], data['store_id'])
+        item = ItemModel(name, **data)
 
         try:
             item.save_to_db()
         except:
-            return {"message": "An error occurred inserting the item."}, 500
+            return {"message": "An error occurred while inserting the item."}, 500
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['isAdmin']:
+            return {'message': 'Admin privilege required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {'message': 'Item deleted'}
+            return {'message': 'Item deleted.'}
+        return {'message': 'Item not found.'}, 404
 
     def put(self, name):
-        data = Item.parser.parse_args()
+        data = self.parser.parse_args()
 
         item = ItemModel.find_by_name(name)
 
         if item:
             item.price = data['price']
         else:
-            item = ItemModel(name, data['price'])
+            item = ItemModel(name, **data)
 
         item.save_to_db()
 
@@ -69,9 +74,10 @@ class Item(Resource):
 
 
 class ItemList(Resource):
-    @jwt_required
+    @jwt_optional
     def get(self):
-        claims = get_jwt_claims()
-        if claims['isAdmin']:
-            return {'items': list(map(lambda x: x.json(), ItemModel.query.all()))}, 200
-        return {'message': 'Admin privilege required'}, 401
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {'items': [item['name'] for item in items]}, 401
